Run unit tests in the same process
